requirements.txt
aiohttp==3.7.4.post0
attrs==21.2.0
chardet==4.0.0
idna==3.2
multidict==5.1.0
pytz==2021.1
yarl==1.6.3
PyYAML~=5.4.1
requests==2.26.0
textblob~=0.15.3
pip~=21.1.2
wheel~=0.36.2
six~=1.16.0
nltk~=3.6.2
click~=8.0.1
tqdm~=4.62.2
regex~=2021.8.21
colorama~=0.4.4
joblib~=1.0.1
certifi~=2021.5.30
urllib3~=1.26.6
pycparser~=2.20
setuptools~=57.0.0